--- src/components/about/Card.tsx
@@ -1,4 +1,0 @@
-const Card = () => {
-  return <div className="">Card</div>;
-};
-export default Card;
