+++ src/components/about/CourseOfferingCard.tsx
@@ -1,0 +1,4 @@
+const CourseOfferingCard = () => {
+  return <div className="">Card</div>;
+};
+export default CourseOfferingCard;
